更新转换器，修复进度条值设置为整数，增强日志记录，优化文件夹创建逻辑

+++ core/convert/__init__.py
@@ -9,7 +9,7 @@ class ConvertWorker(QThread):
     """
 
     # 定义信号：传递当前进度（0-100）
-    progress_updated = pyqtSignal(int)
+    progress_updated = pyqtSignal(float)
     # 定义信号：任务完成（无参数）
     task_finished = pyqtSignal()
     error_occurred = pyqtSignal(str)
